feat:Broke down TeamMembersPage into smaller reusable components
