+++ apps/web/app/team-members/page.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { TeamMembersPage } from '~/components/pages/team-members/team-members'
+
+export default function page() {
+	return (
+		<div>
+			<TeamMembersPage />
+		</div>
+	)
+}
